Refactors FacilityFeatures Content Block
- component does not carry internal state, so we refactor it into a
named function expression for easier debugging in the browser.
- the FeaturesList functional component is not used in another content
block so we move it directly into the FacilityFeatures component.
- we remove the map function and hardcode the list content for now in
order to implement the updated style rules for the project. we keep
the features array for now -- we may want to move back to a map
function to output list contents later